+++ travel-planner/src/pages/Flights/booking.tsx
@@ -1,6 +1,6 @@
 // src/pages/BookingPage.tsx
 import { useEffect, useState } from "react";
-import type { Traveler } from "../../interfaces/Booking";
+import type { Billing, FlightOrder, Traveler } from "../../interfaces/Booking";
 import { FaChevronRight, FaCheck } from "react-icons/fa";
 import { createTravelersFromPricing } from "../../utils/createTravelersFromPricing";
 import { useLocation } from "react-router-dom";
@@ -8,12 +8,77 @@ import PassengerModal from "../../components/PassengerModal";
 import BookingHeader from "../../components/flight/BookingHeader";
 import ItineraryFareDetails from "../../components/flight/ItineraryFareDetails";
 import Button from "../../components/common/Button";
+import type { FlightOffer } from "../../interfaces/ConfirmedFlightOffer";
+import BillingModal from "../../components/flight/BillingModal";
 
 export default function FlightBooking() {
   const [passengers, setPassengers] = useState<Traveler[]>([]);
   const [activePassenger, setActivePassenger] = useState<Traveler | null>(null);
+
+  const billing: Billing = {
+    addresseeName: { firstName: "", middleName: "", lastName: "" },
+    companyName: "N/A",
+    purpose: "STANDARD",
+    phones: [{ deviceType: "MOBILE", countryCallingCode: "263", number: "" }],
+    emailAddress: "",
+    address: {
+      lines: ["", "null"],
+      postalCode: "",
+      cityName: "",
+      countryCode: "",
+    },
+  };
+
+  const [billingData, setBillingData] = useState<Billing>(billing);
+  const [showBillingModal, setShowBillingModal] = useState(false);
+
   const location = useLocation();
-  const flight = location.state?.flight;
+  const flight: FlightOffer = location.state?.flight;
+
+  const billingCompleted =
+    billingData.addresseeName.firstName &&
+    billingData.addresseeName.lastName &&
+    billingData.companyName &&
+    billingData.emailAddress &&
+    billingData.purpose &&
+    billingData.phones.length > 0 &&
+    billingData.address.lines.length > 0 &&
+    billingData.address.cityName &&
+    billingData.address.postalCode &&
+    billingData.address.countryCode;
+
+  const handleConfirmBooking = async () => {
+    if (!billingData || passengers.length === 0) {
+      alert("Please fill in all traveler and billing details.");
+      return;
+    }
+    const order: FlightOrder = {
+      data: {
+        type: "flight-order",
+        flightOffers: [flight], // From search results or itinerary
+        travelers: passengers,
+        contacts: [billingData],
+        remarks: {
+          general: [
+            {
+              subType: "GENERAL_MISCELLANEOUS",
+              text: "Booking created via Voyant",
+            },
+          ],
+        },
+        ticketingAgreement: {
+          option: "DELAY_TO_CANCEL",
+          delay: "6D", // example: 6 days
+        },
+      },
+    };
+    console.log(order);
+  };
+
+  const handleBillingSave = (updatedBilling: Billing) => {
+    setBillingData(updatedBilling);
+    setShowBillingModal(false);
+  };
 
   useEffect(() => {
     if (flight?.travelerPricings?.length) {
@@ -32,13 +97,21 @@ export default function FlightBooking() {
     <div className="max-w-2xl md:max-w-6xl md:flex gap-10 mx-auto p-6 mt-8 space-y-4">
       <BookingHeader flight={flight} />
       <div className="lg:flex-2/3">
+        {/* Passenger Info */}
         <div className="mb-8 p-4 bg-[var(--color-bg-solid)] rounded-2xl border border-[var(--color-border)] shadow hover:shadow-md transition ">
           <h2 className="font-semibold mb-3">Passenger Information</h2>
           {passengers.map((passenger) => {
             const completed =
               passenger.name.firstName &&
               passenger.name.lastName &&
-              passenger.contact.emailAddress;
+              passenger.gender &&
+              passenger.documents[0].number &&
+              passenger.documents[0].issuanceCountry &&
+              passenger.documents[0].issuanceLocation &&
+              passenger.documents[0].nationality &&
+              passenger.documents[0].validityCountry &&
+              passenger.documents[0].issuanceDate &&
+              passenger.documents[0].expiryDate;
             const initials =
               passenger.name.firstName && passenger.name.lastName
                 ? `${passenger.name.firstName[0]}${passenger.name.lastName[0]}`
@@ -85,6 +158,49 @@ export default function FlightBooking() {
           )}
         </div>
 
+        {/* Billing Info*/}
+        <div className="mb-8 p-4 bg-[var(--color-bg-solid)] rounded-2xl border border-[var(--color-border)] shadow hover:shadow-md transition">
+          <h2 className="font-semibold mb-3">Billing Information</h2>
+
+          <div
+            onClick={() => {
+              setShowBillingModal(true);
+            }}
+            className="flex items-center justify-between py-4 cursor-pointer"
+          >
+            <div className="flex items-center space-x-3">
+              <div className="w-10 h-10 flex items-center justify-center rounded-full bg-[var(--color-accent)] font-bold text-[var(--color-text-primary)]">
+                B
+              </div>
+
+              <div>
+                <p className="font-medium text-[var(--color-text-primary)]">
+                  Billing Contact
+                </p>
+                <p className="text-sm text-[var(--color-text-secondary)]">
+                  {billingCompleted
+                    ? "Details completed"
+                    : "Tap to add billing details"}
+                </p>
+              </div>
+            </div>
+
+            {billingCompleted ? (
+              <FaCheck className="text-green-500" />
+            ) : (
+              <FaChevronRight className="text-[var(--color-text-muted)]" />
+            )}
+          </div>
+
+          {showBillingModal && (
+            <BillingModal
+              billing={billingData}
+              onSave={handleBillingSave}
+              onClose={() => setShowBillingModal(false)}
+            />
+          )}
+        </div>
+
         <ItineraryFareDetails travelerPricings={flight.travelerPricings} />
         {/* Disclaimer */}
         <p className="flex gap-2 text-sm text-[var(--color-text-muted)] m-8 max-w-2xl mx-auto">
@@ -99,8 +215,8 @@ export default function FlightBooking() {
         {/* Confirm Button */}
         <Button
           label="Confirm & Pay"
-          type="submit"
           ariaLabel={`Confirm booking `}
+          onClick={handleConfirmBooking}
           className="mt-6 block mx-auto bg-gradient-to-br from-yellow-500 via-amber-600 to-yellow-700 bottom-4 backdrop-blur-4xl sticky"
         />
       </div>